# Avoid unnecessary replacements in action text when the replacement node is unchanged