@jorgemanrubia 2022-12-17
> Is that 30% of overall time or time for just this method call? Either way is a nice improvement, but it would be good to include a benchmark in the commit message to clarify. Also, it might be good to compare different sized replacements in the benchmark.

@jonathanhefner I added a benchmark to the PR description and commit. Sorry that was missing in the first place. I did the test in our app and it was a bit complex to share here, but I prepared a synthetic one that is simpler.

Actually the performance gain from the benchmark is bigger (+50%), although it is obviously dependent on the payload type: the improvement is proportional to the number of nodes to ignore in your replacement logic.