@jonathanhefner 2022-12-18
A couple of things I wanted to note about the benchmark: It's usually good to reduce overhead in benchmark blocks as much as possible so that the comparison is more direct.  For example, (re)defining a method incurs a lot of overhead.  Instead of redefining `replace` on each iteration, we can define `old_replace` and `new_replace` methods, and then call those as appropriate.  Also, 2000 elements seems like a lot.  Is that typical?  I put together a benchmark that has 200 elements instead, but each element has more content (i.e. more expensive `node.to_s`).

```ruby
# frozen_string_literal: true
require "benchmark/ips"

class ActionText::Fragment
  def old_replace(selector)
    update do |source|
      source.css(selector).each do |node|
        node.replace(yield(node).to_s)
      end
    end
  end

  def new_replace(selector)
    update do |source|
      source.css(selector).each do |node|
        replacement_node = yield(node)
        node.replace(replacement_node.to_s) if node != replacement_node
      end
    end
  end
end

LOREM = <<~HTML
  <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor
  incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis
  nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
  Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore
  eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt
  in culpa qui officia deserunt mollit anim id est laborum.</p>
HTML

REPLACE = LOREM.sub("<p>", "<p>REPLACE ")

fragment = ActionText::Fragment.wrap(<<~HTML)
  <div>
    #{LOREM * 100}
    #{REPLACE * 100}
  </div>
HTML

Benchmark.ips do |x|
  x.report "old_replace" do
    fragment.old_replace("p") do |node|
      node.text.start_with?("REPLACE") ? LOREM : node
    end
  end

  x.report "new_replace" do
    fragment.new_replace("p") do |node|
      node.text.start_with?("REPLACE") ? LOREM : node
    end
  end

  x.compare!
end
```

```console
$ bin/rails r 46751.rb
Warming up --------------------------------------
         old_replace     5.000  i/100ms
         new_replace    11.000  i/100ms
Calculating -------------------------------------
         old_replace     59.463  (±21.9%) i/s -    275.000  in   5.117922s
         new_replace    114.447  (±26.2%) i/s -    495.000  in   5.068100s

Comparison:
         new_replace:      114.4 i/s
         old_replace:       59.5 i/s - 1.92x  (± 0.00) slower
```

The performance ratio is pretty similar to yours though, so I'll go ahead and merge.  Thank you, @jorgemanrubia! :wave: